+++ NEWS.md
@@ -11,6 +11,7 @@
 * `environments()` now has it's own `print.mark_environments()` method rather than calling `cat()` within the function itself
 * `array_extract()`'s first argument is changed from `arr` to `.arr`
 * correction to error message in `limit()`
+* adds missing `sort` argument to `props()`
 
 ## New features